airflow/requirements.txt
# Базовые зависимости
pandas>=1.5.0
numpy>=1.22.0
requests>=2.28.0
psycopg2-binary>=2.9.3
python-dotenv>=0.21.0
backoff>=2.2.1
tenacity>=8.2.0
SQLAlchemy>=1.4.0
aiohttp>=3.8.0
pydantic>=1.10.0

# Для работы с файлами
openpyxl>=3.0.10
xlrd>=2.0.1

# Для работы с API маркетплейсов
ozon-seller>=0.1.0
wildberries-api>=0.1.0

# Для логирования
loguru>=0.6.0